refactor: extract OperationsCell into separate component
  - Move MaintenanceRequest class to datamodels_ui.py for better organization
  - Remove explicit session.commit() calls, let SQLAlchemy handle transactions
  - Extract OperationsCell and MaintenanceForm into separate component file
  - Clean up unused imports in WorkerPage.tsx
  - Improve code modularity and maintainability
